Move vm tests
